+++ src/Akeneo/Pim/Structure/Bundle/Resources/workspaces/settings-ui/src/pages/AttributeGroupsIndex.tsx
@@ -7,7 +7,7 @@ import {Breadcrumb} from 'akeneo-design-system';
 const AttributeGroupsIndex: FC = () => {
   const {groups, load, isPending} = useAttributeGroupsIndexState();
   const translate = useTranslate();
-  const settingsHomePageRoute = useRoute('pim_settings_index');
+  const settingsHomePageRoute = `#${useRoute('pim_settings_index')}`;
 
   useEffect(() => {
     (async () => {